requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
diff-match-patch==20230430
Django==5.0.6
django-environ==0.11.2
django-extensions==3.2.3
django-import-export==4.0.1
et-xmlfile==1.1.0
flake8==7.0.0
import-export==0.3.1
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
